dm: move struct passthru_dev to header file
Tracked-On: #5168
